+++ lightyear/src/client/io/mod.rs
@@ -1,0 +1,45 @@
+//! Wrapper around a transport, that can perform additional transformations such as
+//! bandwidth monitoring or compression
+pub(crate) mod config;
+pub(crate) mod transport;
+
+use crate::transport::error::{Error, Result};
+use crate::transport::io::{BaseIo, IoState};
+use async_channel::{Receiver, Sender};
+use bevy::prelude::{Deref, DerefMut, Real, Res, Resource, Time};
+use tokio::sync::mpsc;
+
+pub struct IoContext {
+    pub(crate) event_sender: Option<ClientNetworkEventSender>,
+    pub(crate) event_receiver: Option<ClientIoEventReceiver>,
+}
+
+/// Client IO
+pub type Io = BaseIo<IoContext>;
+
+impl Io {
+    pub fn close(&mut self) -> Result<()> {
+        self.state = IoState::Disconnected;
+        if let Some(event_sender) = self.context.event_sender.as_mut() {
+            event_sender
+                .send_blocking(ClientIoEvent::Disconnected(
+                    std::io::Error::other("client requested disconnection").into(),
+                ))
+                .map_err(Error::from)?;
+        }
+        Ok(())
+    }
+}
+
+/// Events that will be sent from the io thread to the main thread
+/// (so that we can update the netcode state when the io changes)
+pub(crate) enum ClientIoEvent {
+    Connected,
+    Disconnected(Error),
+}
+
+#[derive(Deref, DerefMut)]
+pub(crate) struct ClientIoEventReceiver(pub(crate) Receiver<ClientIoEvent>);
+
+#[derive(Deref, DerefMut)]
+pub(crate) struct ClientNetworkEventSender(pub(crate) Sender<ClientIoEvent>);
