+++ lightyear/src/client/mod.rs
@@ -24,6 +24,7 @@ mod easings;
 #[cfg_attr(docsrs, doc(cfg(feature = "leafwing")))]
 #[cfg(feature = "leafwing")]
 pub mod input_leafwing;
+pub(crate) mod io;
 pub(crate) mod message;
 pub(crate) mod networking;
 pub mod replication;
