+++ lightyear/src/connection/server.rs
@@ -6,8 +6,12 @@ use crate::connection::id::ClientId;
 #[cfg(all(feature = "steam", not(target_family = "wasm")))]
 use crate::connection::steam::server::SteamConfig;
 use crate::packet::packet::Packet;
-use crate::prelude::{Io, IoConfig, LinkConditionerConfig};
+use crate::prelude::client::ClientTransport;
+use crate::prelude::server::ServerTransport;
+use crate::prelude::LinkConditionerConfig;
 use crate::server::config::NetcodeConfig;
+use crate::server::io::Io;
+use crate::transport::config::SharedIoConfig;
 
 pub trait NetServer: Send + Sync {
     /// Start the server
@@ -48,6 +52,8 @@ pub trait NetServer: Send + Sync {
 pub struct ServerConnection {
     server: Box<dyn NetServer>,
 }
+
+pub type IoConfig = SharedIoConfig<ServerTransport>;
 
 /// Configuration for the server connection
 #[derive(Clone, Debug)]
