+++ lightyear/src/connection/steam/client.rs
@@ -2,7 +2,8 @@ use crate::client::networking::NetworkingState;
 use crate::connection::client::NetClient;
 use crate::connection::id::ClientId;
 use crate::packet::packet::Packet;
-use crate::prelude::{Io, LinkConditionerConfig};
+use crate::prelude::client::Io;
+use crate::prelude::LinkConditionerConfig;
 use crate::serialize::bitcode::reader::BufferPool;
 use crate::transport::LOCAL_SOCKET;
 use anyhow::{anyhow, Context, Result};
