+++ lightyear/src/connection/steam/server.rs
@@ -3,8 +3,9 @@ use crate::connection::id::ClientId;
 use crate::connection::netcode::MAX_PACKET_SIZE;
 use crate::connection::server::NetServer;
 use crate::packet::packet::Packet;
-use crate::prelude::{Io, LinkConditionerConfig};
+use crate::prelude::LinkConditionerConfig;
 use crate::serialize::bitcode::reader::BufferPool;
+use crate::server::io::Io;
 use anyhow::{anyhow, Context, Result};
 use bevy::utils::HashMap;
 use std::collections::VecDeque;
