+++ lightyear/src/tests/stepper.rs
@@ -9,9 +9,10 @@ use bevy::MinimalPlugins;
 
 use crate::connection::netcode::generate_key;
 use crate::prelude::client::{
-    Authentication, ClientCommands, ClientConfig, InterpolationConfig, PredictionConfig, SyncConfig,
+    Authentication, ClientCommands, ClientConfig, ClientTransport, InterpolationConfig,
+    PredictionConfig, SyncConfig,
 };
-use crate::prelude::server::{NetcodeConfig, ServerCommands, ServerConfig};
+use crate::prelude::server::{NetcodeConfig, ServerCommands, ServerConfig, ServerTransport};
 use crate::prelude::*;
 use crate::tests::protocol::*;
 
@@ -84,13 +85,13 @@ impl BevyStepper {
         // channels to receive a message from/to server
         let (from_server_send, from_server_recv) = crossbeam_channel::unbounded();
         let (to_server_send, to_server_recv) = crossbeam_channel::unbounded();
-        let client_io = IoConfig::from_transport(ClientTransport::LocalChannel {
+        let client_io = client::IoConfig::from_transport(ClientTransport::LocalChannel {
             send: to_server_send,
             recv: from_server_recv,
         })
         .with_conditioner(conditioner.clone());
 
-        let server_io = IoConfig::from_transport(ClientTransport::Channels {
+        let server_io = server::IoConfig::from_transport(ServerTransport::Channels {
             channels: vec![(addr, to_server_recv, from_server_send)],
         })
         .with_conditioner(conditioner.clone());
