+++ lightyear/src/transport/error.rs
@@ -12,4 +12,30 @@ pub enum Error {
     #[cfg(all(feature = "websocket", not(target_family = "wasm")))]
     #[error(transparent)]
     WebSocket(#[from] tokio_tungstenite::tungstenite::error::Error),
+    #[error("could not send message via channel: {0}")]
+    Channel(String),
+}
+
+#[allow(unused_qualifications)]
+impl<T> ::core::convert::From<async_channel::SendError<T>> for Error {
+    #[allow(deprecated)]
+    fn from(source: async_channel::SendError<T>) -> Self {
+        Error::Channel(source.to_string())
+    }
+}
+
+#[allow(unused_qualifications)]
+impl<T> ::core::convert::From<crossbeam_channel::SendError<T>> for Error {
+    #[allow(deprecated)]
+    fn from(source: crossbeam_channel::SendError<T>) -> Self {
+        Error::Channel(source.to_string())
+    }
+}
+
+#[allow(unused_qualifications)]
+impl<T> ::core::convert::From<tokio::sync::mpsc::error::SendError<T>> for Error {
+    #[allow(deprecated)]
+    fn from(source: tokio::sync::mpsc::error::SendError<T>) -> Self {
+        Error::Channel(source.to_string())
+    }
 }
