+++ lightyear/src/transport/io.rs
@@ -7,12 +7,10 @@ use std::net::{IpAddr, SocketAddr};
 use bevy::app::{App, Plugin};
 use bevy::diagnostic::{Diagnostic, DiagnosticPath, Diagnostics, RegisterDiagnostic};
 use bevy::prelude::{Deref, DerefMut, Real, Res, Resource, Time};
-use crossbeam_channel::Sender;
 #[cfg(feature = "metrics")]
 use metrics;
 use tracing::info;
 
-use crate::transport::local::{LocalChannel, LocalChannelBuilder};
 use crate::transport::middleware::conditioner::{
     ConditionedPacketReceiver, LinkConditioner, LinkConditionerConfig, PacketLinkConditioner,
 };
@@ -20,24 +18,17 @@ use crate::transport::middleware::PacketReceiverWrapper;
 use crate::transport::{PacketReceiver, PacketSender, Transport};
 
 use super::error::{Error, Result};
-use super::{BoxedCloseFn, BoxedReceiver, BoxedSender, LOCAL_SOCKET};
+use super::{BoxedReceiver, BoxedSender};
 
 /// Connected io layer that can send/receive bytes
 #[derive(Resource)]
-pub struct Io {
+pub struct BaseIo<T: Send + Sync> {
     pub(crate) local_addr: SocketAddr,
     pub(crate) sender: BoxedSender,
     pub(crate) receiver: BoxedReceiver,
-    pub(crate) close_fn: Option<BoxedCloseFn>,
     pub(crate) state: IoState,
-    pub(crate) event_receiver: Option<ClientIoEventReceiver>,
     pub(crate) stats: IoStats,
-}
-
-impl Default for Io {
-    fn default() -> Self {
-        panic!("Io::default() is not implemented. Please provide an io");
-    }
+    pub(crate) context: T,
 }
 
 // TODO: add stats/compression to middleware
@@ -49,7 +40,7 @@ pub struct IoStats {
     pub packets_received: usize,
 }
 
-impl Io {
+impl<T: Send + Sync> BaseIo<T> {
     pub fn local_addr(&self) -> SocketAddr {
         self.local_addr
     }
@@ -62,23 +53,15 @@ impl Io {
     pub fn stats(&self) -> &IoStats {
         &self.stats
     }
-
-    pub fn close(&mut self) -> Result<()> {
-        self.state = IoState::Disconnected;
-        if let Some(close_fn) = std::mem::take(&mut self.close_fn) {
-            close_fn()?;
-        }
-        Ok(())
-    }
 }
 
-impl Debug for Io {
+impl<T: Send + Sync> Debug for BaseIo<T> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Io").finish()
     }
 }
 
-impl PacketReceiver for Io {
+impl<T: Send + Sync> PacketReceiver for BaseIo<T> {
     fn recv(&mut self) -> Result<Option<(&mut [u8], SocketAddr)>> {
         // todo: bandwidth monitoring
         self.receiver.as_mut().recv().map(|x| {
@@ -96,7 +79,7 @@ impl PacketReceiver for Io {
     }
 }
 
-impl PacketSender for Io {
+impl<T: Send + Sync> PacketSender for BaseIo<T> {
     fn send(&mut self, payload: &[u8], address: &SocketAddr) -> Result<()> {
         // todo: bandwidth monitoring
         #[cfg(feature = "metrics")]
@@ -178,41 +161,4 @@ pub(crate) enum IoState {
     Connecting,
     Connected,
     Disconnected,
-}
-
-#[derive(Deref, DerefMut)]
-pub(crate) struct ClientIoEventReceiver(Receiver<ClientIoEvent>);
-
-/// Events that will be sent from the io thread to the main thread
-/// (so that we can update the netcode state when the io changes)
-pub(crate) enum ClientIoEvent {
-    Connected,
-    Disconnected(Error),
-}
-
-#[derive(Deref, DerefMut)]
-pub(crate) struct ClientNetworkEventSender(Sender<ClientIoEvent>);
-
-#[derive(Deref, DerefMut)]
-pub(crate) struct ServerIoEventReceiver(Receiver<ServerIoEvent>);
-
-/// Events that will be sent from the io thread to the main thread
-pub(crate) enum ServerIoEvent {
-    ServerConnected,
-    ServerDisconnected(Error),
-    /// the io thread for a given client got disconnected
-    ClientDisconnected(SocketAddr),
-}
-
-#[derive(Deref, DerefMut)]
-pub(crate) struct ServerNetworkEventSender(async_channel::Sender<ServerNetworkEvent>);
-
-/// Event sent from the main thread to the io thread
-/// (usually to close the io thread if the main thread requested the server to stop,
-/// or to close a client's io thread if the client requested a disconnection)
-pub(crate) enum ServerNetworkEvent {
-    /// The server is stopped, we should stop the io thread
-    ServerDisconnected,
-    /// A client id has sent a disconnection packet, we should stop the io thread
-    ClientDisconnected(SocketAddr),
 }
