+++ lightyear/src/transport/websocket/client_native.rs
@@ -27,14 +27,12 @@ use tokio_tungstenite::{
 use tracing::{debug, info, trace};
 use tracing_log::log::error;
 
-use crate::transport::client::{ClientTransportBuilder, ClientTransportEnum};
+use crate::client::io::transport::{ClientTransportBuilder, ClientTransportEnum};
+use crate::client::io::{ClientIoEvent, ClientIoEventReceiver, ClientNetworkEventSender};
 use crate::transport::error::{Error, Result};
-use crate::transport::io::{
-    ClientIoEvent, ClientIoEventReceiver, ClientNetworkEventSender, IoState,
-};
+use crate::transport::io::IoState;
 use crate::transport::{
-    BoxedCloseFn, BoxedReceiver, BoxedSender, PacketReceiver, PacketSender, Transport,
-    LOCAL_SOCKET, MTU,
+    BoxedReceiver, BoxedSender, PacketReceiver, PacketSender, Transport, LOCAL_SOCKET, MTU,
 };
 
 pub(crate) struct WebSocketClientSocketBuilder {
@@ -52,7 +50,7 @@ impl ClientTransportBuilder for WebSocketClientSocketBuilder {
     )> {
         let (serverbound_tx, mut serverbound_rx) = unbounded_channel::<Message>();
         let (clientbound_tx, clientbound_rx) = unbounded_channel::<Message>();
-        let (close_tx, mut close_rx) = crossbeam_channel::bounded(1);
+        let (close_tx, close_rx) = async_channel::bounded(1);
         // channels used to check the status of the io task
         let (status_tx, status_rx) = async_channel::bounded(1);
 
@@ -111,7 +109,7 @@ impl ClientTransportBuilder for WebSocketClientSocketBuilder {
                     }
                 }));
                 // wait for a signal that the io should be closed
-                close_rx.recv().await;
+                let _ = close_rx.recv().await;
                 let _ = status_tx
                     .send(ClientIoEvent::Disconnected(
                         std::io::Error::other("websocket closed").into(),
